requirements.txt
bluepy
awoxpy
paho-mqtt
construct